+++ src/cpp/multimap/internal/Varint.hpp
@@ -49,15 +49,15 @@ struct Varint {
     Limits() = delete;
   };
 
-  static size_t readUint(const char* buffer, size_t size, uint32_t* value);
+  static uint32_t readUint(const char* buffer, size_t size, uint32_t* value);
   // Reads a 32-bit unsigned integer from `buffer` into `value`.
   // Returns the number of bytes read on success, otherwise zero.
   // Preconditions:
   //  * `buffer` is not null
   //  * `value` is not null
 
-  static size_t readUintWithFlag(const char* buffer, size_t size,
-                                 uint32_t* value, bool* flag);
+  static uint32_t readUintWithFlag(const char* buffer, size_t size,
+                                   uint32_t* value, bool* flag);
   // Reads a 32-bit unsigned integer with flag from `buffer` into `value` and
   // `flag`. Returns the number of bytes read on success, otherwise zero.
   // Preconditions:
@@ -65,15 +65,15 @@ struct Varint {
   //  * `value` is not null
   //  * `flag` is not null
 
-  static size_t writeUint(uint32_t value, char* buffer, size_t size);
+  static uint32_t writeUint(uint32_t value, char* buffer, size_t size);
   // Writes a 32-bit unsigned integer into `buffer`.
   // Returns the number of bytes written on success, otherwise zero.
   // Preconditions:
   //  * `value` is not greater than `MAX_N4`
   //  * `buffer` is not null
 
-  static size_t writeUintWithFlag(uint32_t value, bool flag, char* buffer,
-                                  size_t size);
+  static uint32_t writeUintWithFlag(uint32_t value, bool flag, char* buffer,
+                                    size_t size);
   // Writes a 32-bit unsigned integer with flag into `buffer`.
   // Returns the number of bytes written on success, otherwise zero.
   // Preconditions:
